Move shared fixtures and constants to files
